remember to move the bootstrap cookbook
into the dotfiles.org file.

Really, this is one of the weaknesses I find in the entire "literate
programming" model. Debugging something like this can be rather
hard. When you don't know what is wrong, having the discipline to
maintain the literate style is really hard. I'm not even necessarily
sure that it is a good idea to keep doing it like that anyway.

Besides that, I feel like there should be an automated way of
"importing" files into the literate file structure. If I generate a
bunch of files, I ought to be able to easily bring them into a LP
project.