adds toggle sudo nopassword script
The reasons for this needing to exist are complicated.

First, the environment variable SUDO_USER is set occasionally to know
who is being set. however, if sudo is called recursively, SUDO_USER
becomes root (because of the user who called sudo a second time was
already root). So, levels of sudo nesting need to be well-controlled as
long as SUDO_USER is used.

This is all tangential. The real reason I need this to exist is because
sudo is needed occasionally inside my recipes. I suppose the above
explanation pertains to why the recipe itself is not just run using sudo
entirely. Thus, for the duration of the chef convergence, sudoless
password is turned on for the current user.

I *think* I might only ever need this in tests, though.